网关实现

+++ lingo-common/src/main/java/com/moncoder/lingo/common/api/IStatusCode.java
@@ -8,7 +8,7 @@ package com.moncoder.lingo.common.api;
  */
 public interface IStatusCode {
 
-    Long getCode();
+    Integer getCode();
 
     String getMessage();
 }
